// src/app/api/models/notification-client-dto.ts
/* tslint:disable */
export class NotificationClientDto {
  ClientId?: null | number;
  Content?: null | string;
  CreationDate?: null | string;
  ExpiryDate?: null | string;
  ImageUrl?: null | string;
  IsAvailable?: null | boolean;
  IsDelete?: null | boolean;
  IsRead?: null | boolean;
  LastEditDate?: null | string;
  Notes?: null | string;
  NotificationCid?: number;
  Title?: null | string;
}